+++ apps/repl/package.json
@@ -118,7 +118,7 @@
     "typescript": "^5.1.6",
     "vfile": "^6.0.0",
     "webpack": "^5.88.1",
-    "yn": "^4.0.0"
+    "yn": "^5.0.0"
   },
   "engines": {
     "node": ">= v16",
