+++ apps/tutorial/package.json
@@ -90,7 +90,7 @@
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-qunit": "^8.0.0",
     "fs-extra": "^11.1.1",
-    "globby": "^11.1.0",
+    "globby": "^13.2.1",
     "loader.js": "^4.7.0",
     "pnpm-sync-dependencies-meta-injected": "^0.0.8",
     "prettier": "^2.8.8",
