+++ spikes/url-shortening/package.json
@@ -6,7 +6,7 @@
     "@algorithm.ts/base64": "3.1.1",
     "@algorithm.ts/huffman": "3.1.1",
     "filesize": "^10.0.7",
-    "globby": "^13.1.3",
+    "globby": "^13.2.1",
     "jsoncrush": "^1.1.8",
     "lz-string": "^1.5.0",
     "tinybench": "^2.5.0"
